samples/flask/205.summarize-bot/requirements.txt
Flask>=1.0.2
asyncio>=3.4.3
requests>=2.18.1
botframework-connector>=4.4.0.b1
botbuilder-schema>=4.4.0.b1
botbuilder-core>=4.4.0.b1
botbuilder-dialogs>=4.4.0.b1
botbuilder-ai>=4.4.0.b1
certifi
cffi
cchardet
colorama
cycler
cssselect
ekphrasis
ftfy
github3.py
glob2
html5lib
idna
joblib
kiwisolver
lxml
matplotlib
mkl-fft
mkl-random
nltk
numpy
olefile
Pillow
pip
pycparser
pyparsing
python-dateutil
PyYAML
pyzmq
scikit-learn
scipy
setuptools
six
sumy
tabulate
termcolor
tornado
tqdm
urllib3
visdom
wcwidth
websocket-client
wheel
spacy
pytorch-pretrained-bert
gensim
multiprocess
pyrouge
tensorboardX
corenlp
urlextract
wget